Fix `unsafe_wrap` in `MagneticDataset` to let CI pass
See https://github.com/JuliaLang/julia/issues/52308

+++ src/magnetic.jl
@@ -149,7 +149,7 @@ function MagneticDataset(dataset::SpglibMagneticDataset)
                 unsafe_wrap(
                     Matrix{Float64},
                     dataset.std_tensors,
-                    (3, Int64(dataset.n_std_atoms)),  # Issue to Julia community
+                    (3, Int(dataset.n_std_atoms)),  # Issue to Julia community
                 ),
             ),
         )
